% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Cscale.R
\name{add_Cscale}
\alias{add_Cscale}
\title{Color scale}
\usage{
add_Cscale(
  pos = "1/1",
  title = "Depth (m)",
  width = 18,
  height = 70,
  Bcuts = Depth_cuts,
  Bcols = Depth_cols,
  minVal = NA,
  maxVal = NA,
  fontsize = 1,
  offset = 100,
  lwd = 1
)
}
\arguments{
\item{pos}{character, indicating the vertical position of the color scale (which is always on the 
right side of plots). if \code{pos="1/1"}, the color scale will be centered. 
if \code{pos="1/2"}, the color scale will in the top half of the plotting regions.
if \code{pos="2/2"}, the color scale will in the bottom half of the plotting regions.}

\item{title}{character, title of the color scale.}

\item{width}{width of the color scale box, expressed in \% of the width of the plotting region.}

\item{height}{height of the color scale box, expressed in \% of the height of the plotting region.}

\item{Bcuts}{numeric vector of color classes. May be generated via \code{\link{add_col}}.}

\item{Bcols}{character vector of color names. May be generated via \code{\link{add_col}}.}

\item{minVal}{If desired, the color scale may be generated starting from the value \code{minVal}. See examples.}

\item{maxVal}{If desired, the color scale may be generated up to the value \code{maxVal}. See examples.}

\item{fontsize}{Size of the text in the color scale.}

\item{offset}{Controls the horizontal position of the color scale. Increase to distance from the plotting region.}

\item{lwd}{thickness of lines.}
}
\description{
Adds a color scale to plots. Default behavior set for bathymetry.
}
\examples{

Example 1: simple bathymetry plot with color scale
par(mai=c(0,0,0,1)) #plot margins as c(bottom, left, top, right), here giving space for the color scale
plot(SmallBathy, breaks=Depth_cuts, col=Depth_cols, legend=F,axes=F,box=F)
add_Cscale(height=95)

Example 2: Show only values greater than 'minVal'
par(mai=c(0,0,0,1)) #plot margins as c(bottom, left, top, right), here giving space for the color scale
plot(SmallBathy, breaks=Depth_cuts, col=Depth_cols, legend=F,axes=F,box=F)
add_Cscale(minVal=-3200)

Example 3: Show only values between 'minVal' and 'maxVal'
par(mai=c(0,0,0,1)) #plot margins as c(bottom, left, top, right), here giving space for the color scale
plot(SmallBathy, breaks=Depth_cuts, col=Depth_cols, legend=F,axes=F,box=F)
add_Cscale(minVal=-3200,maxVal=-400)

Example 4: Adding two color scales

#Bathymetry
par(mai=c(0,0,0,1)) #plot margins as c(bottom, left, top, right), here giving space for the color scales
plot(SmallBathy,breaks=Depth_cuts,col=Depth_cols,legend=F,axes=F,box=F)
add_Cscale(pos='1/2',height=45,maxVal=-1,minVal=-4000,fontsize=0.8)

#Some gridded data
MyGrid=create_PolyGrids(GridData,dlon=2,dlat=1)
Gridcol=add_col(MyGrid$Catch_sum,cuts=10)
plot(MyGrid,col=Gridcol$varcol,add=T)
#Add color scale using cuts and cols generated by add_col, note the use of 'round'
add_Cscale(pos='2/2',height=45,title='Catch (t)',Bcuts=round(Gridcol$cuts,1),Bcols=Gridcol$cols,fontsize=0.8)

}
\seealso{
\code{\link{SmallBathy}}, \code{\link{Depth_cuts}}, \code{\link{Depth_cols}}, \code{\link{add_col}},
\href{http://www.stat.columbia.edu/~tzheng/files/Rcolor.pdf}{R colors}.
}
