% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clip2Coast.R
\name{Clip2Coast}
\alias{Clip2Coast}
\title{Clip Polygons to the Antarctic coastline}
\usage{
Clip2Coast(Pl, Coastline, ID)
}
\arguments{
\item{Pl}{polygon(s) to be clipped}

\item{Coastline}{SCAR 2016 coastline data see ?Antarctic_coastline (data still needs to be added) for further details, but need to explore low and high res options}

\item{ID}{unique IDs for polygons, but think I will get rid of this as polys should have unique IDs}
}
\description{
Clip Polygons to the Antarctic coastline defined by SCAR in 2016
}
\keyword{Clip}
\keyword{Coastline}
