% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{Coast}
\alias{Coast}
\title{Simplified and subsettable coastline}
\format{
sf
}
\usage{
data(Coast)
}
\description{
Coastline polygons generated from \link{load_Coastline} and sub-sampled to only contain data that falls
within the boundaries of the Convention Area. This spatial object may be subsetted to plot the coastline for selected
ASDs or EEZs (see examples). Source: \url{http://gis.ccamlr.org/}
}
\examples{
#Complete coastline:
plot(st_geometry(Coast[Coast$ID=='All',]),col='grey')

#ASD 48.1 coastline:
plot(st_geometry(Coast[Coast$ID=='48.1',]),col='grey')
}
\seealso{
\code{\link{Clip2Coast}}, \code{\link{load_Coastline}}.
}
