% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS.R
\docType{data}
\name{Coastline_data}
\alias{Coastline_data}
\title{Coastline data}
\format{SpatialPolygonsDataFrame}
\source{
https://gis.ccamlr.org/home see information (i) under Coastlines
}
\description{
Coastline data
}
