% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_RefGrid}
\alias{create_RefGrid}
\title{Create a Reference grid}
\usage{
create_RefGrid(
  bb,
  ResLat = 1,
  ResLon = 2,
  LabLon = NA,
  lwd = 1,
  fonsize = 1,
  offset = NA
)
}
\arguments{
\item{bb}{bounding box of the first plotted object. for example, \code{bb=bbox(SmallBathy)} or \code{bb=bbox(MyPolys)}.}

\item{ResLat}{Latitude resolution in decimal degrees.}

\item{ResLon}{Longitude resolution in decimal degrees.}

\item{LabLon}{Longitude at which Latitude labels should appear. if set, the resulting Reference grid will be circumpolar.}

\item{lwd}{Line thickness of the Reference grid.}

\item{offset}{offset of the Reference grid's labels (distance to plot border).}

\item{fontsize}{Font size of the Reference grid's labels.}
}
\description{
Create a Latitude/Longitude reference grid and add it to maps.
}
\examples{

Example 1: Circumpolar grid with Latitude labels at Longitude 0

par(mai=c(1,1.5,0.5,0)) #Figure margins as c(bottom, left, top, right)
plot(SmallBathy,breaks=Depth_cuts, col=Depth_cols, legend=F,axes=F,box=F)
create_RefGrid(bb=bbox(SmallBathy),ResLat=10,ResLon=20,LabLon = 0)

Example 2: Local grid around created polygons

MyPolys=create_Polys(PolyData,Densify=T)
BathyC=crop(SmallBathy,MyPolys) #crop the bathymetry to match the extent of MyPolys
par(mai=c(0.5,0.5,0.5,0.5)) #Figure margins as c(bottom, left, top, right)
plot(BathyC,breaks=Depth_cuts, col=Depth_cols, legend=F,axes=F,box=F)
create_RefGrid(bb=bbox(BathyC),ResLat=2,ResLon=6)
plot(MyPolys,add=T,col='orange',border='brown',lwd=2)

}
\seealso{
\code{\link{SmallBathy}}
}
