% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensifyData.R
\name{DensifyData}
\alias{DensifyData}
\title{Densify line or polygon data when projecting}
\usage{
DensifyData(Lon, Lat)
}
\arguments{
\item{Lon}{Longitude coordinates (in Decimal Degrees)}

\item{Lat}{Latitude coordinates (in Decimal Degrees)}
}
\description{
A geoprocessing function that densifies a line or polygon data when projecting. This function is called in create_Polys
}
\keyword{Densify}
