% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{Depth_cuts}
\alias{Depth_cuts}
\title{Bathymetry depth classes}
\format{Numeric vector}
\usage{
plot(SmallBathy,breaks=Depth_cuts,col=Depth_cols,axes=F,box=F)
}
\description{
Set of depth classes to plot bathymetry, to be used in conjunction with \link{Depth_cols}.
}
\seealso{
\code{\link{add_cols}}, \code{\link{CBar}}, \code{\link{SmallBathy}}.
}
