% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{Depth_cuts2}
\alias{Depth_cuts2}
\title{Bathymetry depth classes with Fishable Depth range}
\format{
numeric vector
}
\usage{
data(Depth_cuts2)
}
\description{
Set of depth classes to plot bathymetry and highlight Fishable Depth range (600-1800), to be used in conjunction with \link{Depth_cols2}.
}
\examples{
plot(SmallBathy,breaks=Depth_cuts2,col=Depth_cols2,axes=FALSE,box=FALSE)
}
\seealso{
\code{\link{Depth_cuts}}, \code{\link{add_col}}, \code{\link{add_Cscale}}, \code{\link{SmallBathy}}.
}
