% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMStoDD.R
\name{DMSToDD}
\alias{DMSToDD}
\title{DMSToDD}
\usage{
DMSToDD(Coords)
}
\arguments{
\item{Coords}{is a single vector of either Longitude or Latitude values in Degrees, minutes and seconds, but can also handle Degrees and Decimal minutes}
}
\description{
Convert Coordinates that are in Degrees and decimal minutes or Degrees, minutes and seconds to decimal degrees
}
\keyword{Coordinates,}
\keyword{Degrees}
\keyword{Minutes}
\keyword{Seconds}
