% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEZ_data.R
\docType{data}
\name{EEZ_data}
\alias{EEZ_data}
\title{Exclusive Economic Zones (EEZs)}
\format{A Spatial Polygon Data Frame with 4 rows and 1 variable:
\describe{
  \item{EEZ_CODE}{the code and label for an EEZ}
  ...
}}
\source{
\url{https://gis.ccamlr.org/home}
}
\usage{
EEZ_data
}
\description{
Polygon data with attributes generated by the Secretariat in 2015 that defines the boundaries of Exclusive Economic Zones in the Convention Area
}
\keyword{datasets}
