% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{GridData}
\alias{GridData}
\title{Example dataset for create_PolyGrids}
\format{
data.frame
}
\usage{
data(GridData)
}
\description{
To be used in conjunction with \link{create_PolyGrids}.
}
\examples{
#View(GridData)

MyGrid=create_PolyGrids(Input=GridData,dlon=2,dlat=1)
plot(st_geometry(MyGrid),col=MyGrid$Col_Catch_sum)
}
\seealso{
\code{\link{create_PolyGrids}}.
}
