% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ASDs.R
\name{load_ASDs}
\alias{load_ASDs}
\title{Load CCAMLR Statistical Areas and Divisions}
\usage{
load_ASDs(format)
}
\arguments{
\item{format}{"GEOJSON" will extract the latest Statistical Areas and Division geographical reference data from the CCAMLR GIS geoserver and "RDATA" will use the Spatial Polygon Data Frame last saved with the package}
}
\description{
Load CCAMLR Statistical Areas and Divisions (ASDs)
}
\keyword{Areas}
\keyword{Divisions}
\keyword{Statistical}
\keyword{and}
