% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Coastline.R
\name{load_Coastline}
\alias{load_Coastline}
\title{load CCAMLR Coastline}
\usage{
load_Coastline(format)
}
\arguments{
\item{format}{two different formats can be specified. To extract the latest Coastline reference data from the CCAMLR GIS website enter "GEOJSON" and to use the Spatial Polygon Data Frame last saved with the package enter "RDATA" see ?Coastline_data for further details/documentation}
}
\description{
Load CCAMLR Coastline based on SCAR Coastline data
}
\keyword{Coastline}
