% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_EEZs.R
\name{load_EEZs}
\alias{load_EEZs}
\title{Load Exclusive Economic Zones}
\usage{
load_EEZs(format)
}
\arguments{
\item{format}{"RDATA" will use the Spatial Polygon Data Frame last saved with the package see EEZ_data for further details}
}
\description{
Load Exclusive Economic Zones (EEZs) that fall within the Convention Area
}
\keyword{Economic}
\keyword{Exclusive}
\keyword{Zones}
