% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_GEBCO.R
\name{load_GEBCO}
\alias{load_GEBCO}
\title{Load GEBCO bathymetry data}
\usage{
load_GEBCO(data_dir)
}
\arguments{
\item{data_dir}{is the directory into which the most current bathymetry data
(gebco 2014 30-arc second grid) will be downloaded and stored}
}
\description{
Load GEBCO bathymetry data
}
