% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_MAs.R
\name{load_MAs}
\alias{load_MAs}
\title{Load Other Fishery Management Areas}
\usage{
load_MAs(format)
}
\arguments{
\item{format}{"RDATA" will use the Spatial Polygon Data Frame last saved with the package see ?MA_data for documentation of data/further details}
}
\description{
Load Other Fishery Management Areas
}
\keyword{Areas}
\keyword{Fishery}
\keyword{Management}
\keyword{Other}
