% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_MPAs.R
\name{load_MPAs}
\alias{load_MPAs}
\title{load CCAMLR MPAs}
\usage{
load_MPAs(format)
}
\arguments{
\item{format}{two different formats can be specified. To extract the latest MPA geographical reference data from the CCAMLR GIS website enter "GEOJSON" and to use the Spatial Polygon Data Frame last saved with the package enter "RDATA" see ?MPA_data for further details/documentation}
}
\description{
Load CCAMLR Marine Protected Areas (MPAs)
}
\keyword{Areas}
\keyword{Marine}
\keyword{Protected}
