% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_RBs.R
\name{load_RBs}
\alias{load_RBs}
\title{Load CCAMLR Research Blocks}
\usage{
load_RBs(format)
}
\arguments{
\item{format}{"GEOJSON" will extract the latest Research Block geographical reference data from the CCAMLR GIS website and "RDATA" will use the Spatial Polygon Data Frame last saved with the package}
}
\description{
Load CCAMLR Research Blocks (RBs)
}
\keyword{Blocks}
\keyword{Research}
