% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_RefAreas}
\alias{load_RefAreas}
\title{Load CCAMLR Reference Areas}
\usage{
load_RefAreas(format)
}
\arguments{
\item{format}{"GEOJSON" will extract this geographical reference data displayed on the CCAMLR GIS website and "RDATA" will use the Spatial Polygon Data Frame last saved with the package}
}
\description{
Load CCAMLR Reference Areas (RefAreas)
}
\keyword{Reference}
\keyword{areas}
