% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_SSMUs.R
\name{load_SSMUs}
\alias{load_SSMUs}
\title{Load CCAMLR Small Scale Management Units}
\usage{
load_SSMUs(format)
}
\arguments{
\item{format}{"RDATA" will use the Spatial Polygon Data Frame last saved with the package see ?SSMU_data for documentation of data/further details}
}
\description{
Load CCAMLR Small Scale Management Units (SSMUs)
}
\keyword{(SSMUs)}
\keyword{Management}
\keyword{Scale}
\keyword{Small}
\keyword{Units}
