% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_SSRUs.R
\name{load_SSRUs}
\alias{load_SSRUs}
\title{load CCAMLR SSRUs}
\usage{
load_SSRUs(format)
}
\arguments{
\item{format}{two different formats can be specified. To extract the latest SSRU geographical reference data from the CCAMLR GIS website enter "GEOJSON" and to use the Spatial Polygon Data Frame last saved with the package enter "RDATA" see ?SSRU_data for further details/documentation}
}
\description{
Load CCAMLR Small Scale Research Units (SSRUs)
}
\keyword{Research}
\keyword{SSRU}
\keyword{Scale}
\keyword{Small}
\keyword{Units}
