% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS.R, R/MA_data.R
\docType{data}
\name{MA_data}
\alias{MA_data}
\alias{MA_data}
\title{MA data}
\format{SpatialPolygonsDataFrame}
\source{
https://gis.ccamlr.org/home see information (i) under Management areas

\url{https://gis.ccamlr.org/home}
}
\usage{
MA_data
}
\description{
MA data

Polygon data with attributes generated by the Secretariat in 2013 that defines the boundaries of Management Areas in the South Sandwich Islands and South Georgia
}
\keyword{datasets}
