% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{PolyData}
\alias{PolyData}
\title{Example dataset for create_Polys}
\format{
data.frame
}
\usage{
data(PolyData)
}
\description{
To be used in conjunction with \link{create_Polys}.
}
\examples{
#View(PolyData)

MyPolys=create_Polys(PolyData,Densify=TRUE)
plot(st_geometry(MyPolys),col='green')
text(MyPolys$Labx,MyPolys$Laby,MyPolys$ID)
plot(st_geometry(MyPolys[MyPolys$ID=='three',]),border='red',lwd=3,add=TRUE)
}
\seealso{
\code{\link{create_Polys}}.
}
