% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS.R
\docType{data}
\name{RefAreas}
\alias{RefAreas}
\title{RefAreas 
 
 These were Reference Areas used in local toothfish biomass estimation (see WG-FSA-17/42 for further details)}
\format{SpatialPolygonsDataFrame}
\source{
see inst/lrobinson-notes/createRefAreas.R
}
\description{
RefAreas 
 
 These were Reference Areas used in local toothfish biomass estimation (see WG-FSA-17/42 for further details)
}
