% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabed_area.R
\name{seabed_area}
\alias{seabed_area}
\title{Calculate planimetric seabed area}
\usage{
seabed_area(bathymetry, Polys, fishable_area, depth_classes = NA)
}
\arguments{
\item{bathymetry}{is data in a raster file format upon which the calculation
is based, it is assumed that all bathymetry values above zero have been removed.}

\item{Polys}{is the polygon(s) region of interest in the SpatialPolygonsDataFrame}

\item{fishable_area}{is TRUE if the planimetric seabed area in the fishable depth range (i.e. 800-1600m) will be calculated if FALSE then the total seabed area will be calculated}

\item{depth_classes}{is a character vector that includes the minimum and maximum depth within each depth class e.g. c("0-600","600-1800","1800-max")}
}
\description{
Calculate planimetric seabed area within the single depth class and/or multiple depth classes
}
