% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS.R, R/SSMU_data.R
\docType{data}
\name{SSMU_data}
\alias{SSMU_data}
\alias{SSMU_data}
\title{SSMU data}
\format{SpatialPolygonsDataFrame}
\source{
https://gis.ccamlr.org/home see information (i) under Management areas

\url{https://gis.ccamlr.org/home}
}
\usage{
SSMU_data
}
\description{
SSMU data

Polygon data with attributes generated by the Secretariat in 2013 that defines the boundaries of Small Scale Management Units in the Convention Area
}
\keyword{datasets}
