% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS.R, R/SSRU_data.R
\docType{data}
\name{SSRU_data}
\alias{SSRU_data}
\alias{SSRU_data}
\title{SSRU data}
\format{SpatialPolygonsDataFrame}
\source{
https://gis.ccamlr.org/home see information (i) under Management areas

\url{https://gis.ccamlr.org/home}
}
\usage{
SSRU_data
}
\description{
SSRU data

Polygon data with attributes generated by the Secretariat in 2016 that defines the boundaries of Small Scale Research Units in the Convention Area
}
\keyword{datasets}
