% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seabed_area.R
\name{sum_area}
\alias{sum_area}
\title{Sum planimetric seabed area}
\usage{
sum_area(bathymetry)
}
\arguments{
\item{bathymetry}{is the bathymetry raster data upon which the planimetric seabed area calculation
is based}
}
\description{
Sum planimetric seabed area
}
