% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{Add labels}
\usage{
add_labels(layer = NULL, fontsize = 1, fonttype = 1, angle = 0, col = "black")
}
\arguments{
\item{layer}{single or vector of characters, may only be one, some or all of: 
\code{c("ASDs","SSRUs","RBs","SSMUs","MAs","RefAreas","MPAs","EEZs")}.}

\item{fontsize}{size of the text.}

\item{fonttype}{type of the text (1 to 4), where 1 corresponds to plain text, 
2 to bold face, 3 to italic and 4 to bold italic.}

\item{angle}{rotation of the text in degrees.}

\item{col}{color of the text.}
}
\value{
Adds labels to plot.
}
\description{
Adds labels to plots of the layers obtained via 'load_' functions. Positions correspond to the centroids
of polygon parts. Internally used in conjunction with \code{\link{Labels}}.
To manually generate your own labels, see \code{\link[rgeos]{polygonsLabel}}.
}
\examples{

#label ASDs in bold and red

ASDs=load_ASDs()
plot(ASDs)
add_labels('ASDs',fontsize=1,fonttype=2,col='red')

#add MPAs and EEZs and their labels in large, green and vertical text
MPAs=load_MPAs()
EEZs=load_EEZs()
plot(MPAs,add=TRUE,border='green')
plot(EEZs,add=TRUE,border='green')
add_labels(c('EEZs','MPAs'),fontsize=2,col='green',angle=90)

}
\seealso{


\code{\link{Labels}}, \code{\link{load_ASDs}}, \code{\link{load_SSRUs}}, \code{\link{load_RBs}},
\code{\link{load_SSMUs}}, \code{\link{load_MAs}}, \code{\link{load_EEZs}},
\code{\link{load_RefAreas}}, \code{\link{load_MPAs}}, \code{\link[rgeos]{polygonsLabel}}.
}
