% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS.R
\docType{package}
\name{CCAMLRGIS}
\alias{CCAMLRGIS}
\title{CCAMLRGIS: A package to load and create spatial data, including layers and tools that are relevant to CCAMLR activities.}
\description{
The CCAMLRGIS package provides two broad categories of functions: load functions and create functions.
}
\section{Load functions}{

load functions are used to import CCAMLR geo-referenced layers and include:
load_ASDs, load_SSRUs, load_RBs, load_SSMUs, load_MAs, load_Coastline, load_RefAreas, load_MPAs, and load_EEZs.
}

\section{Create functions}{

create functions are used to create geo-referenced layers from user data and include:
create_Points, create_Lines, create_Polys, create_PolyGrids, create_Stations and create_RefGrid
}

\section{Vignette}{

To learn more about CCAMLRGIS, start with the vignette:
\code{browseVignettes(package = "CCAMLRGIS")}
}

