% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{CCAMLRp}
\alias{CCAMLRp}
\title{CCAMLRGIS Projection}
\format{
Character string
}
\usage{
data(CCAMLRp)
}
\value{
"+proj=laea +lat_0=-90 +lon_0=0 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
}
\description{
The CCAMLRGIS package uses the \href{http://en.wikipedia.org/wiki/Lambert_azimuthal_equal-area_projection}{Lambert azimuthal equal-area projection}.
Source: \url{http://gis.ccamlr.org/}
}
