% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{CCAMLRp}
\alias{CCAMLRp}
\title{CCAMLRGIS Projection}
\format{
Character string
}
\usage{
data(CCAMLRp)
}
\value{
"+proj=laea +lat_0=-90 +lon_0=0 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs"
}
\description{
The CCAMLRGIS package uses the \href{https://en.wikipedia.org/wiki/Lambert_azimuthal_equal-area_projection}{Lambert azimuthal equal-area projection}.
Source: \url{http://gis.ccamlr.org/}.
In order to align with recent developments within Geographic Information Software, this projection
will be accessed via EPSG code \href{https://epsg.org/crs_6932/WGS-84-NSIDC-EASE-Grid-2-0-South.html}{6932}.
}
