% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clip2Coast.R
\name{Clip2Coast}
\alias{Clip2Coast}
\title{Clip Polygons to the Antarctic coastline}
\usage{
Clip2Coast(Pl, Coastline, ID)
}
\arguments{
\item{Pl}{polygon(s) to be clipped}

\item{Coastline}{calls CCAMLR Coastline_data stored in the package}

\item{ID}{unique IDs for polygons, but think I will get rid of this as polys should have unique IDs}
}
\description{
Clip Polygons to the Antarctic CCAMLR Coastline data
}
\keyword{Clip}
\keyword{Coastline}
