% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{Coast}
\alias{Coast}
\title{Coast}
\format{SpatialPolygons}
\source{
\url{https://gis.ccamlr.org/}
}
\usage{
plot(Coast,col='grey')
}
\description{
Coastline polygons generated from \link{load_Coastline} and sub-sampled to only contain data that falls
within the CCAMLR boundaries.
}
\seealso{
\code{\link{Clip2Coast}}.
}
