% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_Points}
\alias{create_Points}
\title{Create Points}
\usage{
create_Points(
  Input,
  NamesIn = NULL,
  Buffer = 0,
  Clip = FALSE,
  SeparateBuf = TRUE
)
}
\arguments{
\item{Input}{input dataframe.

If \code{NamesIn} is not provided, the columns in the \code{Input} must be in the following order:

Latitude, Longitude, Variable 1, Variable 2, ... Variable x.}

\item{NamesIn}{character vector of length 2 specifying the column names of Latitude and Longitude fields in
the \code{Input}. Latitudes name must be given first, e.g.:

\code{NamesIn=c('MyLatitudes','MyLongitudes')}.}

\item{Buffer}{numeric, radius in nautical miles by which to expand the points. Can be specified for
each point (as a numeric vector).}

\item{Clip}{logical, if set to TRUE, polygon parts (from buffered points) that fall on land are removed (see \link{Clip2Coast}).}

\item{SeparateBuf}{logical, if set to FALSE when adding a \code{Buffer},
all spatial objects are merged, resulting in a single spatial object.}
}
\value{
Spatial object in your environment.
Data within the resulting spatial object contains the data provided in the \code{Input} plus
additional "x" and "y" columns which corresponds to the projected points locations 
and may be used to label points (see examples).

To see the data contained in your spatial object, type: \code{View(MyPoints)}.
}
\description{
Create Points to display point locations. Buffering points may be used to produce bubble charts.
}
\examples{
\donttest{

# For more examples, see:
# https://github.com/ccamlr/CCAMLRGIS#create-points

#Simple points with labels

MyPoints=create_Points(Input=PointData)
plot(st_geometry(MyPoints))
text(MyPoints$x,MyPoints$y,MyPoints$name,adj=c(0.5,-0.5),xpd=TRUE)



}

}
\seealso{
\code{\link{create_Lines}}, \code{\link{create_Polys}}, \code{\link{create_PolyGrids}},
\code{\link{create_Stations}}, \code{\link{create_Pies}}, \code{\link{add_Legend}}.
}
