% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_PolyGrids}
\alias{create_PolyGrids}
\title{Create a Polygon Grid}
\usage{
create_PolyGrids(
  Input,
  OutputFormat = "ROBJECT",
  OutputName = NULL,
  dlon,
  dlat
)
}
\arguments{
\item{Input}{the name of the \code{Input} data as a .csv file or an R dataframe.
If a .csv file is used as input, this file must be in your working directory and its name given in quotes
e.g. "DataFile.csv".

\strong{The columns in the \code{Input} must be in the following order:

Latitude, Longitude, Variable 1, Variable 2 ... Variable x.}}

\item{OutputFormat}{can be an R object or an ESRI Shapefile. if \code{OutputFormat} is specified as
"ROBJECT" (the default), a SpatialPolygonDataFrame is created in your R environment.
if \code{OutputFormat} is specified as "SHAPEFILE", an ESRI Shapefile is exported in
your working directory.}

\item{OutputName}{if \code{OutputFormat} is specified as "SHAPEFILE", the name of the output
shapefile in quotes (e.g. "MyGrid") must be provided.}

\item{dlon}{width of the grid cells in decimal degrees of longitude.}

\item{dlat}{height of the grid cells in decimal degrees of latitude.}

\item{Area}{Area, in square kilometers, of the grid cells.}

\item{cuts}{Number of desired color classes.}

\item{cols}{Desired colors.}
}
\value{
Spatial object in your environment or ESRI shapefile in your working directory.
Data within the resulting object contains the data provided in the \code{Input} after aggregation
within cells. For each Variable, the minimum, maximum, mean, sum, count, standard deviation, and, 
median of values in each cell is returned (see View(MyGrid@data)).

In addition, colors are generated for each aggregatted values according the chosen \code{cuts} and \code{cols}.
To generate a color scale, refer to \link{CBar}.
}
\description{
Create a Polygon Grid to spatially aggregate data in cells of chosen size.
Cell size may be specified in degrees or as a desired area in square kilometers
(in which case all cells are of equal area).
}
