% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_PolyGrids}
\alias{create_PolyGrids}
\title{Create a Polygon Grid that is compatible with CCAMLR online GIS}
\usage{
create_PolyGrids(InputFile, OutputFormat = "ROBJECT", OutputName = NULL,
  dlon, dlat, Type)
}
\arguments{
\item{InputFile}{the name of the input data file in quotes e.g. "DataFile.csv" or an R dataframe}

\item{OutputFormat}{is "ROBJECT" is the default and returns as SpatialPolygonDataFrame  to your R work enviornment and "SHAPEFILE" will write an ESRI Shapefile}

\item{OutputName}{if "SHAPEFILE" format is specified then supply the name of the output shapefile in quotes e.g."MyShape", the default is NULL and assumes an "ROBJECT" format}

\item{dlon}{width of the grid cells in decimal degrees of longitude e.g. dlon=1}

\item{dlat}{height of the grid cells in decimal degrees of latitude e.g.: dlat=0.5}

\item{Type}{set according to the type of data to be gridded. If it is numerical data (e.g. Krill length, tootfish catch; see Table 4) set Type="NUM". If it is categorical data (e.g. CCAMLR Species Code, Vessel name; see Table 5), set Type="CAT".}
}
\description{
Create a Polygon Grid that is compatible with CCAMLR online GIS
}
