% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_Polys}
\alias{create_Polys}
\title{Create Polygons that are compatible with CCAMLR online GIS}
\usage{
create_Polys(InputFile, OutputFormat = "ROBJECT", OutputName = NULL,
  Buffer = 0, Separate = 0, Densify = 1, Clip = 0)
}
\arguments{
\item{InputFile}{the name of the input data file in quotes e.g. "DataFile.csv" or an R dataframe with column values in the following order: Poly Name, Lats and Lons}

\item{OutputFormat}{is "ROBJECT" is the default and returns as SpatialPolygonDataFrame  to your R work enviornment and "SHAPEFILE" will write an ESRI Shapefile}

\item{OutputName}{if "SHAPEFILE" format is specified then supply the name of the output shapefile in quotes e.g."MyShape", the default is NULL and assumes an "ROBJECT" format}

\item{Buffer}{is the value in nautical miles to apply to the line. The default value is 0, assuming no Buffer}

\item{Separate}{currently, if the data file contains several items (e.g. several fishing lines) you may wish to produce one shapefile per item to display them separately on a map. In such a case, set Separate to 1. I want to remove this as I think you can manage this at the input end.}

\item{Densify}{is set to 1 as a default, which will add additional points between points of equal latitude when data are projected. If set to 0 then no additional points will be added}

\item{Clip}{"Coast_Low" will clip a polygon that intersect with the coastline to remove the land and keep only the ocean area, "Coast_Medium" is a higher resolution coastline is also provided with the package, the default is set to 0 which assumes no clipping is required}
}
\description{
Create Polygons that are compatible with CCAMLR online GIS
}
\examples{
## specify the name of the Research block

Name <-"5841_6"

## specify the Longitude coordinates in decimal degrees

Lons <- c(130, 130, 134, 134)

## specify the Latitude coordinates in decimal degrees

Lats <- c(-64.0, -65.5, -65.5,-64.0)

## bind information together into a dataframe 

Coords <- data.frame(Name=rep(Name,length(Lons)),Lats=Lats,Lons=Lons)

## create polygon of proposed Research area

New_RBs <-create_Polys(Coords)
}
