% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{Depth_cols}
\alias{Depth_cols}
\title{Bathymetry colors}
\format{Character vector}
\usage{
data(Depth_cols)
}
\description{
Set of standard colors to plot bathymetry, to be used in conjunction with \link{Depth_cuts}.
}
\examples{
plot(SmallBathy,breaks=Depth_cuts,col=Depth_cols,axes=FALSE,box=FALSE)
}
\seealso{
\code{\link{add_col}}, \code{\link{add_Cscale}}, \code{\link{SmallBathy}}.
}
