% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{Depth_cuts}
\alias{Depth_cuts}
\title{Bathymetry depth classes}
\format{
numeric vector
}
\usage{
data(Depth_cuts)
}
\description{
Set of depth classes to plot bathymetry, to be used in conjunction with \link{Depth_cols}.
}
\examples{
terra::plot(SmallBathy(),breaks=Depth_cuts,col=Depth_cols,axes=FALSE,box=FALSE)
}
\seealso{
\code{\link{Depth_cuts2}}, \code{\link{add_col}}, \code{\link{add_Cscale}}, \code{\link{SmallBathy}}.
}
