% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{Labels}
\alias{Labels}
\title{Polygon labels}
\format{dataframe}
\description{
Labels for the layers obtained via 'load_' functions. Positions correspond to the centroids
of polygon parts. to be used in conjunction with \code{\link{add_labels}}.
}
\examples{
\donttest{


#View(Labels)

ASDs=load_ASDs()
plot(ASDs)
add_labels(mode='auto',layer='ASDs',fontsize=1,fonttype=2)

}

}
\seealso{
\code{\link{add_labels}}, \code{\link{load_ASDs}}, \code{\link{load_SSRUs}}, \code{\link{load_RBs}},
\code{\link{load_SSMUs}}, \code{\link{load_MAs}}, \code{\link{load_EEZs}},
\code{\link{load_RefAreas}}, \code{\link{load_MPAs}}.
}
