% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{LineData}
\alias{LineData}
\title{Example dataset for create_Lines}
\format{DataFrame}
\usage{
View(LineData)

MyLines=create_Lines(LineData)
plot(MyLines,lwd=2)
}
\description{
To be used in conjunction with \link{create_Lines}.
}
