% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{LineData}
\alias{LineData}
\title{Example dataset for create_Lines}
\format{
DataFrame
}
\usage{
data(LineData)
}
\description{
To be used in conjunction with \link{create_Lines}.
}
\examples{
 
#View(LineData)

MyLines=create_Lines(LineData)
plot(MyLines,lwd=2)
}
\seealso{
\code{\link{create_Lines}}.
}
