% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_ASDs}
\alias{load_ASDs}
\title{Load CCAMLR statistical Areas, Subareas and Divisions}
\usage{
load_ASDs()
}
\description{
Download the up-to-date spatial layer from the online CCAMLRGIS (\url{http://gis.ccamlr.org/}) and load it to your environment.
See examples for offline use. All layers use the Lambert azimuthal equal-area projection
 (\code{\link{CCAMLRp}})
}
\examples{
 
\donttest{

#When online:
ASDs=load_ASDs()
plot(st_geometry(ASDs))

#For offline use, see:
#https://github.com/ccamlr/CCAMLRGIS#32-offline-use


}
}
\seealso{
\code{\link{load_SSRUs}}, \code{\link{load_RBs}},
\code{\link{load_SSMUs}}, \code{\link{load_MAs}}, \code{\link{load_Coastline}},
\code{\link{load_MPAs}}, \code{\link{load_EEZs}}.
}
