% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_ASDs}
\alias{load_ASDs}
\title{Load CCAMLR Statistical Areas and Divisions}
\usage{
load_ASDs(format)
}
\arguments{
\item{format}{"GEOJSON" will extract this geographical reference data displayed on the CCAMLR GIS website and "RDATA" will use the Spatial Polygon Data Frame last saved with the package}
}
\description{
Load CCAMLR Statistical Areas and Divisions (ASDs)
}
\examples{
 
# if online
\dontrun{
  ASDs <- load_ASDs("GEOJSON")
}
# if offline 
ASDs <- load_ASDs("RDATA")
}
\keyword{Areas}
\keyword{Divisions}
\keyword{Statistical}
\keyword{and}
