% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_Bathy}
\alias{load_Bathy}
\title{Load Bathymetry data}
\usage{
load_Bathy(LocalFile, Res = 5000)
}
\arguments{
\item{LocalFile}{To download the data, set to \code{FALSE}. To re-use a downloaded file, set to the full path of the file 
(e.g., \code{LocalFile="C:/Desktop/GEBCO2024_5000.tif"}).}

\item{Res}{Desired resolution in meters. May only be one of: 500, 1000, 2500 or 5000.}
}
\value{
Bathymetry raster.
}
\description{
Download the up-to-date projected GEBCO data from the online CCAMLRGIS (\url{https://gis.ccamlr.org/}) and load it to your environment.
This functions can be used in two steps, to first download the data and then use it. If you keep the downloaded data, you can then
re-use it in all your scripts.
}
\details{
To download the data, you must either have set your working directory using \code{\link[base]{setwd}}, or be working within an Rproject.
In any case, your file will be downloaded to the folder path given by \code{\link[base]{getwd}}.

It is strongly recommended to first download the lowest resolution data (set \code{Res=5000}) to ensure it is working as expected.

To re-use the downloaded data, you must provide the full path to that file, for example:

\code{LocalFile="C:/Desktop/GEBCO2024_5000.tif"}.

This data was reprojected from the original GEBCO Grid after cropping at 40 degrees South. Projection was made using the Lambert
azimuthal equal-area projection (\code{\link{CCAMLRp}}),
and the data was aggregated at several resolutions.
}
\examples{
 
\donttest{

#The examples below are commented. To test, remove the '#'.

##Download the data. It will go in the folder given by getwd():
#Bathy=load_Bathy(LocalFile = FALSE,Res=5000)
#plot(Bathy, breaks=Depth_cuts,col=Depth_cols,axes=FALSE)

##Re-use the downloaded data (provided it's here: "C:/Desktop/GEBCO2024_5000.tif"):
#Bathy=load_Bathy(LocalFile = "C:/Desktop/GEBCO2024_5000.tif")
#plot(Bathy, breaks=Depth_cuts,col=Depth_cols,axes=FALSE)

}

}
\references{
GEBCO Compilation Group (2024) GEBCO 2024 Grid (doi:10.5285/1c44ce99-0a0d-5f4f-e063-7086abc0ea0f)
}
\seealso{
\code{\link{add_col}}, \code{\link{add_Cscale}}, \code{\link{Depth_cols}}, \code{\link{Depth_cuts}},
\code{\link{Depth_cols2}}, \code{\link{Depth_cuts2}}, \code{\link{get_depths}},
\code{\link{create_Stations}}, \code{\link{get_iso_polys}},
\code{\link{SmallBathy}}.
}
