% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_Coastline}
\alias{load_Coastline}
\title{load CCAMLR Coastline}
\usage{
load_Coastline(format)
}
\arguments{
\item{format}{"RDATA" will use the SpatialPolygonsDataFrame last saved with the package}
}
\description{
Load CCAMLR Coastline based on SCAR Coastline data
}
\examples{
 

# if offline 
Coast <- load_Coastline("RDATA")
}
\keyword{Coastline}
