% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_Coastline}
\alias{load_Coastline}
\title{load the full CCAMLR Coastline}
\usage{
load_Coastline(p4s = "+proj=laea +lat_0=-90 +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs")
}
\arguments{
  \item{p4s}{Coordinate reference system in proj4string representation.}
}
\description{
Download the up-to-date spatial layer from the \href{http://gis.ccamlr.org/}{online CCAMLRGIS} and load it to your environment.
The layer's Metadata is accessible by clicking on the red 'i' in the list of layers available on the \href{http://gis.ccamlr.org/}{online CCAMLRGIS}.
See examples for offline use. Note that this coastline expands further north than \link{Coast}.
}
\examples{
 
\donttest{

#When online:
Coastline=load_Coastline()
plot(Coastline)

#If going offline in the future: load and save as RData when online,
# then reload RData when offline:
Coastline=load_Coastline()
EEZs=load_EEZs()
save(list=c('Coastline','EEZs'), file = file.path(tempdir(), "CCAMLRLayers.RData"))
rm(Coastline,EEZs)
load(file.path(tempdir(), "CCAMLRLayers.RData"))

}

}
\seealso{
\code{\link{load_ASDs}}, \code{\link{load_SSRUs}}, \code{\link{load_RBs}},
\code{\link{load_SSMUs}}, \code{\link{load_MAs}},
\code{\link{load_RefAreas}}, \code{\link{load_MPAs}}, \code{\link{load_EEZs}}.
}
