% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_Coastline}
\alias{load_Coastline}
\title{load CCAMLR Coastline}
\usage{
load_Coastline(format)
}
\arguments{
\item{format}{"GEOJSON" will extract this geographical reference data displayed on the CCAMLR GIS website and "RDATA" will use the Spatial Polygon Data Frame last saved with the package}
}
\description{
Load CCAMLR Coastline based on SCAR Coastline data
}
\examples{
 
# if online
Coast <- load_Coastline("GEOJSON")

# if offline 
Coast <- load_Coastline("RDATA")
}
\keyword{Coastline}
