% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_MAs}
\alias{load_MAs}
\title{Load Other Fishery Management Areas
Load Other Fishery Management Areas}
\usage{
load_MAs(format)
}
\arguments{
\item{format}{"GEOJSON" will extract this geographical reference data displayed on the CCAMLR GIS website and "RDATA" will use the Spatial Polygon Data Frame last saved with the package}
}
\description{
Load Other Fishery Management Areas
Load Other Fishery Management Areas
}
\examples{
 
# if online
\dontrun{
MAs <- load_MAs("GEOJSON")
}
# if offline 
MAs <- load_MAs("RDATA")
}
