% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_MPAs}
\alias{load_MPAs}
\title{load CCAMLR MPAs}
\usage{
load_MPAs(format)
}
\arguments{
\item{format}{"GEOJSON" will extract this geographical reference data displayed on the CCAMLR GIS website and "RDATA" will use the Spatial Polygon Data Frame last saved with the package}
}
\description{
Load CCAMLR Marine Protected Areas (MPAs)
}
\keyword{Areas}
\keyword{Marine}
\keyword{Protected}
