% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_MPAs}
\alias{load_MPAs}
\title{load CCAMLR Marine Protected Areas}
\usage{
load_MPAs()
}
\description{
Download the up-to-date spatial layer from the \href{http://gis.ccamlr.org/}{online CCAMLRGIS} and load it to your environment.
The layer's Metadata is accessible by clicking on the red 'i' in the list of layers available on the \href{http://gis.ccamlr.org/}{online CCAMLRGIS}.
See examples for offline use.
}
\examples{
 
\dontrun{

#When online:
MPAs=load_MPAs()
plot(MPAs)

#If going offline in the future: load and save as RData when online,
# then reload RData when offline:
MPAs=load_MPAs()
EEZs=load_EEZs()
save(list=c('MPAs','EEZs'), file = "CCAMLRLayers.RData")
rm(MPAs,EEZs)
load("CCAMLRLayers.RData")

}

}
\seealso{
\code{\link{load_ASDs}}, \code{\link{load_SSRUs}}, \code{\link{load_RBs}},
\code{\link{load_SSMUs}}, \code{\link{load_MAs}}, \code{\link{load_Coastline}},
\code{\link{load_RefAreas}}, \code{\link{load_EEZs}}.
}
