% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_RBs}
\alias{load_RBs}
\title{Load CCAMLR Research Blocks}
\usage{
load_RBs()
}
\description{
Download the up-to-date spatial layer from the \href{https://gis.ccamlr.org/}{online CCAMLRGIS} and load it to your environment.
The layer's Metadata is accessible by clicking on the red 'i' in the list of layers available on the \href{https://gis.ccamlr.org/}{online CCAMLRGIS}.
See examples for offline use.
}
\examples{
 
#When online:
RBs=load_RBs()
plot(RBs)

#If going offline in the future: load and save as RData when online,
# then reload RData when offline:
RBs=load_RBs()
EEZs=load_EEZs()
save(list=c('RBs','EEZs'), file = "CCAMLRLayers.RData")
rm(RBs,EEZs)
load("CCAMLRLayers.RData")

}
\seealso{
\code{\link{load_ASDs}}, \code{\link{load_SSRUs}},
\code{\link{load_SSMUs}}, \code{\link{load_MAs}}, \code{\link{load_Coastline}},
\code{\link{load_RefAreas}}, \code{\link{load_MPAs}}, \code{\link{load_EEZs}}.
}
