% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_RefAreas}
\alias{load_RefAreas}
\title{Load CCAMLR Reference Areas}
\usage{
load_RefAreas(format)
}
\arguments{
\item{format}{"RDATA" will use the SpatialPolygonsDataFrame last saved with the package}
}
\description{
Load CCAMLR Reference Areas (RefAreas)
}
\examples{
 
RefAreas <- load_RefAreas("RDATA")
}
\keyword{Reference}
\keyword{areas}
