% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_SSMUs}
\alias{load_SSMUs}
\title{Load CCAMLR Small Scale Management Units}
\usage{
load_SSMUs()
}
\description{
Download the up-to-date spatial layer from the \href{https://gis.ccamlr.org/}{online CCAMLRGIS} and load it to your environment.
See examples for offline use.
}
\examples{
 
#When online:
SSMUs=load_SSMUs()
plot(SSMUs)

#If going offline in the future: load and save as RData when online - then reload RData when offline:
SSMUs=load_SSMUs()
EEZs=load_EEZs()
save(list=c('SSMUs','EEZs'), file = "CCAMLRLayers.RData")
rm(SSMUs,EEZs)
load("CCAMLRLayers.RData")

}
\seealso{
\code{\link{load_ASDs}}, \code{\link{load_SSRUs}}, \code{\link{load_RBs}},
\code{\link{load_MAs}}, \code{\link{load_Coastline}},
\code{\link{load_RefAreas}}, \code{\link{load_MPAs}}, \code{\link{load_EEZs}}.
}
