% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_SSRUs}
\alias{load_SSRUs}
\title{load CCAMLR Small Scale Research Units}
\usage{
load_SSRUs()
}
\description{
Download the up-to-date spatial layer from the \href{http://gis.ccamlr.org/}{online CCAMLRGIS} and load it to your environment.
The layer's Metadata is accessible by clicking on the red 'i' in the list of layers available on the \href{http://gis.ccamlr.org/}{online CCAMLRGIS}.
See examples for offline use.
}
\examples{
 
\donttest{

#When online:
SSRUs=load_SSRUs()
plot(SSRUs)

#If going offline in the future: load and save as RData when online,
# then reload RData when offline:
SSRUs=load_SSRUs()
EEZs=load_EEZs()
save(list=c('SSRUs','EEZs'), file = "MyFolder/CCAMLRLayers.RData")
rm(SSRUs,EEZs)
load("MyFolder/CCAMLRLayers.RData")

}

}
\seealso{
\code{\link{load_ASDs}}, \code{\link{load_RBs}},
\code{\link{load_SSMUs}}, \code{\link{load_MAs}}, \code{\link{load_Coastline}},
\code{\link{load_RefAreas}}, \code{\link{load_MPAs}}, \code{\link{load_EEZs}}.
}
