% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_data.R
\docType{data}
\name{MA_data}
\alias{MA_data}
\title{Other Fishery Management Areas}
\format{A Spatial Polygon Data Frame with 4 rows and 8 variables:
\describe{
  \item{Id}{not unique and completely useless}
  \item{Areakm2}{probably calculated using gebco 2008 data}
  \item{Toothfish}{whether there is toothfish fishing in this area}
  \item{Icefish}{whether there is icefish fishing in this area}
  \item{MAlabel}{Management area label}
  \item{MAname}{Management area full name}
  \item{MAcode}{Management area code}
  ...
}}
\source{
\url{https://gis.ccamlr.org/home}
}
\usage{
MA_data
}
\description{
Polygon data with attributes generated by the Secretariat in 2013 that defines the boundaries of Management Areas in the South Sandwich Islands and South Georgia
}
\keyword{datasets}
