% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{PolyData}
\alias{PolyData}
\title{Example dataset for create_Polys}
\format{DataFrame}
\usage{
View(PolyData)

MyPolys=create_Polys(PolyData,Densify=T)
plot(MyPolys,col='green',add=T)
text(MyPolys$Labx,MyPolys$Laby,MyPolys$ID)
plot(MyPolys[MyPolys$ID=='three',],border='red',lwd=3,add=T)
}
\description{
To be used in conjunction with \link{create_Polys}.
}
\seealso{
\code{\link{create_Polys}}.
}
