% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS_DataDescription.R
\docType{data}
\name{SmallBathy}
\alias{SmallBathy}
\title{Small bathymetry dataset}
\format{raster}
\usage{
plot(SmallBathy,breaks=Depth_cuts,col=Depth_cols,axes=F,box=F)
}
\description{
Bathymetry dataset derived from the \href{https://www.gebco.net/}{GEBCO 2019} dataset.
Subsampled using raster's \link[raster]{resample} function, using the nearest neighbor method
and a 2500m resolution. Projected using the CCAMLR standard projection (\code{\link{CCAMLRp}}).
\strong{To be only used for large scale illustrative purposes}.
}
\seealso{
\code{\link{add_col}}, \code{\link{add_Cscale}}, \code{\link{Depth_cols}}, \code{\link{Depth_cuts}},
\code{\link{get_depths}}, \code{\link{create_Stations}}.
}
