% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS.R
\docType{data}
\name{SSMU_data}
\alias{SSMU_data}
\title{SSMU data}
\format{SpatialPolygonsDataFrame}
\source{
https://gis.ccamlr.org/home see information (i) under Management areas
}
\description{
SSMU data
}
