% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMU_data.R
\docType{data}
\name{SSMU_data}
\alias{SSMU_data}
\title{Small Scale Managament Unit (SSMU)}
\format{A Spatial Polygon Data Frame with 17 rows and 6 variables:
\describe{
  \item{Id}{not unique and completely useless}
  \item{Areakm2}{probably calculated using gebco 2008 data}
  \item{SSMUcode}{Small Scale Management Unit code}
  \item{SSMUlabel}{Small Scale Management Unit label}
  \item{ASDcode}{Statistical Area or Division Code}
  \item{SSMUname}{Small Scale Management Unit full name}
  ...
}}
\source{
\url{https://gis.ccamlr.org/home}
}
\usage{
SSMU_data
}
\description{
Polygon data with attributes generated by the Secretariat in 2013 that defines the boundaries of Small Scale Management Units in the Convention Area
}
\keyword{datasets}
