% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS.R
\docType{data}
\name{SSRU_data}
\alias{SSRU_data}
\title{SSRU data}
\format{SpatialPolygonsDataFrame}
\source{
https://gis.ccamlr.org/home see information (i) under Management areas
}
\description{
SSRU data
}
