% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCAMLRGIS.R, R/SSRU_data.R
\docType{data}
\name{SSRU_data}
\alias{SSRU_data}
\alias{SSRU_data}
\title{SSRU data
 
A SpatialPolygonsDataFrame.}
\format{A Spatial Polygon Data Frame with 63 rows and 14 variables:
\describe{
  \item{GAR_ID}{Geographical Area Identifier in SQL Server?}
  \item{GAR_Name}{Geographical Area name in SQL Server}
  \item{GAR_Short_Label}{Geographical Area label with no spaces or full stops}
  \item{GAR_Long_Label}{Geographical Area long label with spaces and full stops}
  \item{GAR_Start_Date}{Date created? but has a "Z" saved in the day}
  \item{GAR_End_Date}{all NA values so can we remove?}
  \item{GAR_Description}{Original text description of boundaries taken from a CCAMLR document}
  \item{GAR_Reference}{CCAMLR document where the text description of the boundaries was taken from}
  \item{GAR_Parent_ID}{Not sure what this is}
  \item{ATY_ID}{Area Type Identifier}
  \item{CreatedBy}{The person who created it}
  \item{CreatedOn}{The date it was created}
  \item{ModifiedBy}{The person it was last modified by}
  \item{ModifiedOn}{The date it was last modified}
  ...
}}
\source{
\url{https://gis.ccamlr.org/home}
}
\usage{
SSRU_data
}
\description{
SSRU data
 
A SpatialPolygonsDataFrame.

Polygon data with attributes generated by the Secretariat in 2016 that defines the boundaries of Small Scale Research Units in the Convention Area
}
\keyword{datasets}
